+++ browse/build.gradle
@@ -45,9 +45,6 @@ dependencies {
     implementation libs.androidx.coordinatorlayout
     implementation libs.androidx.swiperefreshlayout
     implementation libs.androidx.navigation.fragment
-    implementation 'androidx.appcompat:appcompat:1.4.2'
-    implementation 'com.google.android.material:material:1.4.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
     coreLibraryDesugaring libs.android.desugar
