+++ browse/src/main/kotlin/com/alfresco/content/browse/tasks/detail/TaskDetailViewModel.kt
@@ -11,6 +11,9 @@ import com.alfresco.content.data.TaskRepository
 import com.alfresco.coroutines.asFlow
 import kotlinx.coroutines.launch
 
+/**
+ * Marked as TaskDetailViewModel class
+ */
 class TaskDetailViewModel(
     state: TaskDetailViewState,
     val context: Context,
