+++ build.gradle
@@ -15,7 +15,6 @@ buildscript {
         classpath libs.spotless
         classpath libs.gradleVersionsPlugin
         classpath libs.gradleLicensePlugin
-        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.0'
 
     }
 }
