+++ component/src/main/java/com/alfresco/content/component/ComponentType.kt
@@ -1,5 +1,8 @@
 package com.alfresco.content.component
 
+/**
+ * Marked as ComponentType class
+ */
 enum class ComponentType(val value: String) {
 
     TEXT("text"),
